requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==5.0.0
billiard==4.2.1
boto3==1.35.90
botocore==1.35.90
celery==5.4.0
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
distlib==0.3.9
Django==5.0.1
django-celery-results==2.5.1
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.14.0
filelock==3.16.1
idna==3.10
Jinja2==3.1.4
jinja2-time==0.2.0
jmespath==1.0.1
kombu==5.4.2
make==0.1.6.post2
MarkupSafe==3.0.2
pillow==11.1.0
pipenv==2023.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg-binary==3.2.3
psycopg-pool==3.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.28.1
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
